Refactor the project cracker into a new file et al. (#618)
* Move some code from BuildCommand.fs to a new file.
Do not report nonexistent substitution parameters (see changes in the param function of ProjectCracker.fs).

* Fix and document the --parameters command line option.

Seq.pairwise [1;2;3;4] returns [1,2; 2,3; 3,4], not [1,2; 3,4].

* Remove a try-with block around File.Exists (it doesn't throw).

And use Process.Start with UseShellExecute on all platforms (it is implemented on .NET Core).

* Use System.Text.Json to serialize the documentation search index.

Remove duplicate code in a helper function in BuildCommand.fs.

* Update packages.

* Make some optimizations in the project cracker.

The most important is the replacement of the tree-based FSharpMap with the hashtable-based readOnlyDict.

* Exit without leaving a stacktrace if --parameters was incorrectly defined.

* Do not catch and then print exceptions at the entry point; let them be thrown.

It is not needed. They will already be reported to the console and AggregateExceptions are not correctly handled.